Full refactoring of the library
- The parser has been rewritten to be faster.
- XML Configuration files parsers has been rewritten too.
- Display of extensions (like extended-arrows or big brackets) has been improved
- Introduce a parser for mhchem stuff
- Fix several crappy bugs

+++ jlatexmath/src/main/java/org/scilab/forge/jlatexmath/StrikeThroughAtom.java
@@ -1,4 +1,4 @@
-/* SmashedAtom.java
+/* StrikeThroughAtom.java
  * =========================================================================
  * This file is part of the JLaTeXMath Library - http://forge.scilab.org/jlatexmath
  *
@@ -59,13 +59,13 @@ public class StrikeThroughAtom extends Atom {
     public Box createBox(TeXEnvironment env) {
         TeXFont tf = env.getTeXFont();
         int style = env.getStyle();
-        float axis = tf.getAxisHeight(style);
-        float drt = tf.getDefaultRuleThickness(style);
+        double axis = tf.getAxisHeight(style);
+        double drt = tf.getDefaultRuleThickness(style);
         Box b = at.createBox(env);
         HorizontalRule rule = new HorizontalRule(drt, b.getWidth(), -axis + drt, false);
         HorizontalBox hb = new HorizontalBox();
         hb.add(b);
-        hb.add(new StrutBox(-b.getWidth(), 0, 0, 0));
+        hb.add(new StrutBox(-b.getWidth(), 0., 0., 0.));
         hb.add(rule);
 
         return hb;
